requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-community
streamlit
